writer_identification/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.0
opencv-python>=4.5.0
albumentations>=1.0.0
pillow>=8.0.0
matplotlib>=3.3.0
scikit-learn>=0.24.0
grad-cam>=1.4.0